Add files via upload
This code describes a simple game that can be played between two players. Players receive seven cards each, one for each round. If the value of the first player's card is higher than that of the other player's card for a given round, they win that round. The player who wins the most rounds wins the game.
This was my attempt for an interim assessment for the "IT Specialist Python" eCollege course.