Upgraded version of the original "card busters"
Upgraded version of the code for the original "card busters" file. Sections of code have been assigned to functions play_game() and final_result(), with an extra function future_enhancements() created for future changes. Docstrings and comments have also been added to the functions. This was my attempt for Interim Assignment 4 of the "IT Specialist Python" course provided by eCollege, Ireland.